fix: use thumb cover in admin panel
